audit: verify + correct the final decompile-vs-source pass (58 files)
Investigated the mass audit edit hunk-by-hunk against the Ghidra decompile (28 files
deep-verified; ~16 cosmetic accepted; sugoroku draw deprioritized). The audit was
high-quality: it fixed many serious latent bugs that compiled clean but were wrong at
runtime, e.g.:
  - CAComponent audio output 44100->32000 Hz + format flags 0xc2c (was 0xc)
  - AepLyrCtrl play/stop/reset state machine (wrong state values, int-into-float play head)
  - PlayTask fed judge prevX/prevY instead of x/y + wrong back-tap delta
  - AcNoteMng arcade note-type read from input+5 (was +4)
  - NoteMng MARK/END fields, updateLongNote pool-slot addressing, makeNote structure
  - PlayResultTask results-screen SE cues; neTextTexture premultiply+half-size glyphs
  - neRenderer clip in local space; AudioManager SE-id packing; DownloadMain rank arrays
  - unsigned range checks (MusicData diff, UserSettingData seVolume), deleteAll reset+save

CORRECTED one real regression the audit introduced:
  - NoteMng::makeNote had width/height swapped (baseX must be screenWidth = DAT_00187b78,
    per LayoutedGLView GetFrontBufferWidth). Kept the agent's correct structural fix
    (6-byte NEON pairing + separate targetX/Y), restored the axes.
Also fixed a pre-existing bug: updateLongNote delta startTick->endTick (0x523c).
One minor imperfection noted (AppDelegate acNotePause gated on gLaunchedFromPush as a
proxy for the binary's separate DAT_00173e72). Structural gate: 0 delimiter issues,
frontier unchanged. Full verdict log in HANDOFF.md.

+++ Project/Game/Data/Save/UserSettingData.mm
@@ -574,7 +574,7 @@ static int neSugorokuTouchSoundBit(int mainMapId) {
 + (short)seVolume {
     int v = [self getInt:@"SeVolume"];
     short r = 0x7f;
-    if (v < 0x7f) r = (short)v;
+    if ((short)v < 0x7f) r = (short)v;
     if (r < 0)    r = 0;
     return r;
 }
@@ -589,7 +589,7 @@ static int neSugorokuTouchSoundBit(int mainMapId) {
 + (short)touchSoundVolume {
     int v = [self getInt:@"TouchSoundVolume"];
     short r = 0x7f;
-    if (v < 0x7f) r = (short)v;
+    if ((short)v < 0x7f) r = (short)v;
     if (r < 0)    r = 0;
     return r;
 }
